api/requirements-prod.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# Vector Search
pgvector==0.3.6

# Redis & Workers
redis==5.2.0
arq==0.26.1

# Embeddings & AI
sentence-transformers==3.3.1
torch==2.5.1
transformers==4.47.1

# Cryptography
PyNaCl==1.5.0
blake3==0.4.1
argon2-cffi==23.1.0
mnemonic==0.21

# Document Processing
PyPDF2==3.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.28.1

# MCP
mcp==1.2.0
anthropic==0.42.0

# QR Codes
qrcode[pil]==8.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1